test(agents): preserve reply fixture types and split terminal coverage
